corrections after review

+++ src/main/java/ru/practicum/shareit/item/ItemServiceImpl.java
@@ -2,6 +2,7 @@ package ru.practicum.shareit.item;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import ru.practicum.shareit.exception.NotFoundException;
 import ru.practicum.shareit.item.model.Item;
 
 import java.util.List;
@@ -13,6 +14,9 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public Item getItemById(Long itemId) {
+        if (itemId == null || itemRepository.getItemById(itemId) == null) {
+            throw new NotFoundException("Item not found");
+        }
         return itemRepository.getItemById(itemId);
     }
 
@@ -28,16 +32,37 @@ public class ItemServiceImpl implements ItemService {
 
     @Override
     public Item addItem(Item item) {
+        if (item == null) {
+            throw new NotFoundException("Item not found");
+        }
+
         return itemRepository.addItem(item);
     }
 
     @Override
     public Item updateItem(Item item) {
+        if (item == null) {
+            throw new NotFoundException("Item not found");
+        }
+
+        if (itemRepository.getItemById(item.getId()) == null) {
+            throw new NotFoundException("Item not found in a map");
+        }
+
         return itemRepository.updateItem(item);
     }
 
     @Override
     public void deleteItem(Long itemId) {
+        Item item = getItemById(itemId);
+
+        if (item == null) {
+            throw new NotFoundException("Item not found");
+        }
+
+        if (itemRepository.getItemById(item.getId()) == null) {
+            throw new NotFoundException("Item not found in a map");
+        }
         itemRepository.deleteItem(itemId);
     }
 }
